+++ bugsnag-plugin-android-anr/src/main/jni/bugsnag_anr.c
@@ -8,18 +8,14 @@ extern "C" {
 #endif
 
 JNIEXPORT void JNICALL Java_com_bugsnag_android_AnrPlugin_enableAnrReporting(
-        JNIEnv *env, jobject _this, jobject anr_buffer) {
-    // only install if provided a buffer
-    if (!(*env)->IsSameObject(env, anr_buffer, NULL)) {
-        bsg_handler_install_anr((*env)->GetDirectBufferAddress(env, anr_buffer));
-    }
+        JNIEnv *env, jobject _this, jboolean callPreviousSigquitHandler) {
+    bsg_handler_install_anr(env, _this, callPreviousSigquitHandler);
 }
 
 JNIEXPORT void JNICALL Java_com_bugsnag_android_AnrPlugin_disableAnrReporting(
         JNIEnv *env, jobject _this) {
     bsg_handler_uninstall_anr();
 }
-
 
 #ifdef __cplusplus
 }